+++ R/geom.R
@@ -220,6 +220,14 @@ setMethod("intersect", signature(x="SpatExtent", y="SpatVector"),
 	}
 )
 
+setMethod("mask", signature(x="SpatVector", mask="SpatVector"), 
+	function(x, mask, inverse=FALSE) { 
+		x@ptr <- x@ptr$mask(mask@ptr, inverse)
+		messages(x, "intersect")
+	}
+)
+
+
 #setMethod("intersect", signature(x="SpatRaster", y="SpatRaster"),
 #	function(x, y) {
 #		a <- crop(x, y)